"Add assessment" step 1
The problem that needs to be further solved is how to perform DOM manipulation in node.js. I try to install. jsdom library and call it. But always showing some problems. I will debug again in due3. Maybe try other methods, because calling the jsdom library is the most common method I found